requirements.txt
python-dateutil~=2.8.1
chardet~=3.0.4
terminaltables~=3.1.0